// frontend/src/types/Pelanggan.ts
export type Pelanggan = {
  _id: string
  id_pelanggan: string
  id_outlet: string
  type_kontak: string
  id_harga: string
  nama: string
  perusahaan: string
  rt: string
  provinsi: string
  kabupaten: string
  kecamatan: string
  desa: string
  email: string
  hp: string
  pengenal: string
  no_pengenal: string
  npwp: string
  koordinat: string
  tanggal: string
}